Enabling support for cascading select
Fix for #126 #127 #129

Its noticed that for fields that are of type cascadingselect
Jira’s schema responds with its type as “array”.
This causes the current implementation to try and 
change the payload for the field as a JSONArray but
Jira fails to deserialize this at its end because it
is not expecting a payload of type array.

The catch here is to sometimes have the metadata validation
disabled and send the payload as constructed by the user
directly as is to Jira. This pull request attempts at enabling
this capability.